requirements.txt
youtube_transcript_api
openai
python-dotenv
deepmultilingualpunctuation

